kern: Separate temporary syscalls and kernel messages in separate modules.
This change completes the base support for kernel modules, as now externally
compiled modules can define their own kernel_server demux and syscall_demux
without any explicit linking from the kernel.

A special section is used for both.

This defines an environment where the kernel core defines the kernel objects
and their interaction, while modules defines the actual user interface for
this.

The only syscalls that will be allowed in the kernel core are the ones
necessary to send messages (msgbuf, replyport, sendmsg and recvmsg) and
those to locate the process when it starts, thread_self etc.

All the rest (for example, the MACH3 interface) will be done in separate
modules, allowing ability to suport existing Mach systems, but also create
new interfaces altogether on a system that runs on the same principles.

The only outstanding work to do is to create an interface for kernel modules
to interact with the kernel core. But it's too early now for fixed interfaces.