mig: Add -serverstub and -serverprefix arguments.
The "-serverstub" argument creates stub implementations of the server
routine implementations, thus allowing to compile a server mig file
when not all routines are implemented. The stub routines are weak
symbols, so they'll be discarded once an implementation is done.

The "-serverprefix" argument is equivalent to the 'serverprefix'
mig command, but defined from command line rather than from the
defs file.