kern: Initial work for pager interface.
This commit moves the zero-page filling part of the vmobj fault to
a pager request interface, and adds a trivial default pager (no
page-out and page-in).

A change here is that we're removing an optimization: in previous
code when we were requesting a page from the pager and the mapping
was roshared, but we were operating on a write fault, we would first
get the page and then push, thus saving a page fault.

In the spirit of correct-first optimize-later, for now this
optimization has been removed, at least until we have a fully working
pager interface.