kern: Make pagefault atomic wrt swapping.
The VM system of MACHINA is really a game between two players: the
VM Objects (that request fault in) and the Memory Controller that
reqests swap outs.
Between these two there's the Memory Cache, that handles pages,
and the Cache Object, that maps object to pages and pagers.

On fault-in, we have to check the cache object to understand what
to do (fault the page in, unshare, or just fixup the mappings),
and then ask the memory cache to do such operation.

But between the checking the cache object and the memory map
operation there might be a swapout operation on that page.

This patch allows the fault path to lock the page while we are doing
a fault in operation on a object offset that has a page mapped, so
that we can do the lookup and the operation atomically with respect
to other operations on that page, specifically the swap ou operation.